Move MediaSearch code into its own extension
This patch introduces code from the WikibaseMediaInfo extension
that comprises the MediaSearch user interface, including the special
page, a server-rendered version of the UI, and a Vue app. All code
represented here has been previously reviewed and merged as part of
the WBMI extension, with the following changes:

- Config variable names changed to remove references to WBMI
- Changed CSS class names to swap out the extension prefix, using
  sdms (Structured Data Media Search) instead of wbmi, and removed
  redundant media-search text (e.g. the old wbmi-media-search-input
  class is now just sdms-input)
- Changed Vue base components to use sd (Structured Data) as a
  prefix for components and CSS class names instead of wbmi
- Split LESS variables into 2 files, one for the extension and one
  for base components (which will, ideally, eventually be pulled
  out into a separate library)
- The WikibaseCirrusSearch extension has been made optional. The only
thing MediaSearch relies on this extension for is the license filter
functionality, so tha filter won't appear if WBCS isn't loaded.

To test locally, enable the module and visit Special:NewMediaSearch.

Bug: T265939
Change-Id: I753c7b569afdbcf14ab4a89f30e58d825932295e
Depends-On: I38cb733b7494097e1f1d8e5ac40a667bda557a17