Fix eventlogging-related dependencies
The mediawiki.user (part of core so can require right away)
module is required for the `mw.user.getPageviewToken()` call.
The ext.eventLogging module is lazy-loaded because it
requires another extension which may or may not be present.
If it doesn't exist, logging will be a no-op instead of
causing a failure.

Bug: T277917
Change-Id: I55071a9a39d490ae79e625cc1e3c497015b71c74