Error handling improvements in JS and PHP
Addresses 2 bugs in the way erroneous input is handled.

Bug #1:

Addresses a bug that caused the UI to wind up in a weird state when the
user arrives on the page with a "type" parameter that is not included
in the list of valid types (ex. "bitmap" instead of "image").

PHP: Updates the getType method in SpecialMediaSearch to check that the
specified "type" param in the user's URL is included in the array
of supported types (tabs). If not, the activeType is set to "image"
when the page is rendered in PHP (url is left unchanged here).

JS: Adds a new beforeCreate hook to App.vue where we can reconcile
the "activeTab" and the user's URL if necessary - before any other app
state is inferred from the URL values. Finally, the url params are silently
updated to correct values using history.replaceState.

Bug #2:

Re-introduces server-side display of the error message when bad search
input is provided (ex: "+ - = && || > < ! ( ) { } [ ] ^ " ~ * ? : \ /").

Updates the search method in SpecialMediaSearch to throw an MWException
if the search results come back containing any error data, and updates
the SERPWidget template to show an error message that matches the Vue.js
error component. Now users should see the same error message regardless
of whether JS is enabled or not.

Bug: T280425
Bug: T281173
Change-Id: I2da96673696405b01c1ef0e721a7c948235b5d60