Move ensureArray helper function to its own file
This function was originally a workaround for the variation in
results format depending on which API was used to retrieve them:
local results arrive in array format, while production results
(used for local development) are formatted as an object.

Since we rely heavily on production results for local development,
this function should probably stick around. This patch moves it to
its own file to sequester it from the actual app code.

Bug: T279098
Change-Id: Ibf5f5848dce0481d9e869f3011ca2ce3137bf42a