Replace MustacheDomTemplateParser with default parser
MustacheDomTemplateParser was a specific implementation
that we needed to support templates that allowed html/DOM
objects to be assigned to the template. Mostly because
we needed to be able to inject OOUI nodes into templates
in JS, but we had to be able to use these same templates
in PHP so...

In this repo, with the current code, we don't have those
constraints & there's no need to use this specific
template parser. Might as well use the simpler original.

Bug: T265939
Change-Id: Iacf63c471ccf977d32278f6f7e90f0629cf94093