Show loading indicator while QuickView data is fetched
When a result is clicked, the following will happen:
- The QuickView aside panel will open immediately
- The window will be scrolled to the top of the panel
- If necessary, search results will be scrolled so that the active
result is in view
- A loading indicator will be displayed while QuickView data is fetched

Once data is loaded, it will appear in the panel. If the user has
QuickView open and navigates to a new result, the loading indicator
will be displayed again while data loads, but only if the process
takes more than half a second. This way, those with fast connections
shouldn't notice the loading process.

Bug: T279098
Change-Id: Ie646f3bfce5100c2bf2a30b03ad583cd6333961d