Restore page state from localstorage on back navigation, if necessary
When a user returns to a previous search session via "back" navigation,
we want to restore the UI to the exact state it was in when they left.

In modern browsers that provide a back/forward cache ("bfcache"), this
is all handled automatically. The in-memory JS application is preserved
and can be picked up at exactly the point where the user left off.
Components are still initialized, all ajax-loaded results are present,
quickview state is preserved, etc. This behavior is supported in current
versions of Safari and Firefox, and should come to Chrome soon.

Browsers that don't support bfcache will fall back to the http cache.
In this case, users who back-navigate to a previous search session will
only see the server-rendered results, and the JS UI will need to
re-initialize itself. Scroll position, subsequently fetched results, and
quickview state all get lost, resulting in an annoying experience.

In these cases (when bfcache is not present) we want to manually stash
and restore the relevant page state in localstorage. The page JS will
still have to re-initialize, but once it does we will restore the
previous session (results, scroll position, quickview state, etc).

We only want to provide this behavior manually if bfcache is not
supported (the browser-provided experience is still better & faster),
so some logic has been added to determine 1) whether the user is
back-navigating and 2) whether bfcache is supported.

Bug: T284021
Change-Id: Ibd46adea235e76894fd7d7dae0d90bdc844516e0