Ensure focus is restored when Quickview is closed via keyboard
The Quickview "close" button had two different inline event handlers
in its template: a @click handler and a @keyup.enter handler. The
idea was to distinguish between mouse clicks and keyboard interactions
for this particular element. If users were interacting via keyboard,
an additional bit of data was included in the emitted "close" event
which let the parent SearchResults component know to programatically
restore focus back to the given search result in the grid.

This used to work but now it doesn't for some reason. It seems like
the @click handler is firing before the @keyup handler, meaning the
additional data isn't getting included.

This patch reworks the even handling logic on this button; now there
is only a single @click handler. The corresponding "close" method
now checks the event.details property - for keyboard-initiated events,
this will always be zero. In that case, the additional data can be
sent to tell the parent to restore focus.

Bug: T286833
Change-Id: Ifdb13f57a386ce71d0bd3cc43f49fad30de9e2c9