Check for null "searchOptions" when restoring filter state
The "onPopstate" handler (used to restore page state when the user
navigates via the "back" button) throws the following error with
some frequency on mobile devices:

TypeError: searchOptions[this.currentTab] is not an Object. (evaluating
'key in searchOptions[this.currentTab]')

Adding a simple check for null before testing object properties
should prevent this.

Bug: T283864
Change-Id: I6db8de56bcd5498da320baa9fa8465027f59a175