Simplify Select component by removing support for unused "items" formats
The Select component code contains (erroneous) logic for alternate types
of items which are never actually used in MediaSearch. Since this
component is not going to be re-used anywhere else (a new shared library
is being developed by the Design Systems Team which we will eventually
migrate this feature to), we should just simplify it and remove the
code for unsed formats.

Bug: T285582
Change-Id: I3fbdd762195dee30884a09a2892f984bf6065343