Ensure that filter state is stashed in mw.storage
Bug T291291 identifies a situation where a user who is relying on
restored page state (after back navigation) can encounter an
inconsistency between results in a hidden tab vs. the current
active filters in that tab.

The reason for this issue is that we were not storing the Vuex
filter state (all active filters across all tabs) in the page
data that gets restored from mw.storage; instead we had assumed
that the URL parameters would be sufficient to restore this.

Unfortunately this is true only for the tab that is active
immediately before the user departs Special:MediaSearch. If they
have other filters active in other tabs, those won't be restored
(unless the user is in a browser that has b/f cache enabled and
preserves the in-memory application in its entirety).

Stashing and then restoring the Vuex filter state for all tabs
fixes this problem.

Bug: T291291
Change-Id: Idee77190d47d7649e527ffc95ba2e550495d2caa