Don't call $refs methods inline in template
Fixes a bug where the namespace filter dialog appears as empty,
because the Radio and Checkbox components inside it were failing
to render. This failure was due to the fact that the templates
of both components contained an inline handler method like this:

    @keydown.prevent.enter="$refs.label.click"

This worked fine in Vue 2, but in Vue 3 compatability mode this
no longer functions apparently. Moving the code into a dedicated
handler method for each component seems to fix the problem.

Bug: T297664
Change-Id: Ic6772dc3421455eca8c4287a48071d35c77fd4be