Don't attempt to scroll to a non-existing result
Adds a guard clause to the SearchResults component's
scrollIntoViewIfNeeded method. When the filters change while
the quickview is open, it is possible for the original result
which triggered the quickview to no longer be present on the page.
This change simply adds a guard clause to check for this scenario
before attempting to scroll back to a result on the page. If the
original result ref is not found, nothing else needs to be done.

Change-Id: Icea7dc12dd579dfb4ad98b0e00ddae3155c36c40
(cherry picked from commit 6a1201b4f02007157239d6877896fd04be0696b4)