Add a dependency on "intersectionObserver-polyfill"
The patch for T300666 will allow modules designated as ES6-only
to run on desktop/mobile Safari versions back to 11.1. Safari 11
supports most of the ES6 standard but lacks support for
IntersectionObserver (which is used in MediaSearch).

This change adds a dependency on the new intersectionObserver-polyfill
module to ensure that the oldest supported Safari users are able to
get full functionality.

Bug: T300666
Depends-On: I1cb370603b2ac13bcaf74890841ec941961810ef
Change-Id: I258c93fc1f1e8ce3b5d626381cf9fc7ddc0c309a