maintenance: Reduce use of DBConnRef and global MediaWikiServices
The DBConnRef class should mostly be a detail that is invisible
to callers, prefering the IDatabase interface instead.
The getConnectionRef() and getConnection() method now both return
this and raw connections are no longer encouraged.

Prefer the provided Maintenance->getDB() method over calling into
global MediaWikiServices. Internally this method does actually do
the same since maintenance scripts are rarely re-used for testability
or cross-wiki invocation, but this interface certainly allows for
injection since the Maintenance class can carry that responsibility,
and we already invoke it with sufficient indirection to allow for
injection to take place (that is, we don't instantiate our the
Maintenance directly, instad using $maintClass to let core invoke
it after any other setup/injection).

This script is unlikely to be maintained much further I imagine, so
this change is mainly intended to make the code more in line with
other maintenance scripts, for awareness of maintainers, and to be
a better starting example when discovered/found/copied for another
purpose in the future.

Change-Id: I03f2f5616ec5f7829de158291d2d50af7a5fd3de