Remove local copy of i18n plugin from jest tests
Previously we had included a local copy of the i18n plugin
from MediaWiki to use during front-end unit tests. This is not
a great long-term strategy because the upstream plugin might
change, and those changes would then need to be re-copied here.

A better solution is to simply mock the plugin during testing.
Vue Test Utils provides a striaghtforward way to do this via
the `config.global` object -- this allows mocking not only methods
but also template directives like `v-i18n-html`.

The relevant code has been added to jest.setup.js, and tests have
been updated accordingly.

This patch also updates the Jest unit tests to use destructured
"require" statements for VueTestUtils, to provide a better experience
in some editors and to make it more obvious which parts of the
VueTestUtils API are being used.

Change-Id: I85d7eed8d824cc9ea9e43ffc6734296516e5b4ce