Update no-JS UI to match new CSS-only Tabs implementation
Codex v0.16.0 made some changes to the markup and styles of
the Tabs component. The MediaSearch SERPWidget template is relying
on the Tabs styles to ship a visually-compatible, no-JS layout
to the client.

This change updates the markup used in the SERPWidget to match
what the udpated Codex Tabs styles expect, to ensure that there
is no user-facing change on this page.

Bug: T340132
Change-Id: I545d789f0d292298c8089386193e8fd8ccea9a4d