MediaSearch filters can wrap on narrow screens.
This is designed to horizontally scroll, but that is just ugly and
people can't even find the functionality. When scrolling is applied,
there is only a gradient to indicate this, the line above at least
has an arrow.

It is also broken in Vector-2022, because content in grids will
overflow the grid boundaries.

Instead, simply wrap this, which looks just fine.

Bug: T334198
Change-Id: Ic9a665b0d778a186fff06d6977634277ed15f1cc