SpecialMediaSearch: Add error handling for search API errors
SpecialMediaSearch is executing the search API via internal PHP code
which ApiMain detects and makes it to bypass the normal API error
handling mechanism with the expectations that the caller would handle
that.

Not handling the API errors means that any error, even simple errors
like invalid parameter name, throws Exception directly in the UI and
logs in error logs as application error.

This patch fixes that by adding error handling for the relevant
ApiUsageException to resolve numerous phabricator tasks.

ApiUsageException will now be caught. The error message (which comes
from the original fatal status by the search API) will then be shown
to the user. The messages are already user messages with proper i18n
strings and because the may contain links, they're parsed as html for
proper rendering.

Also add minimal regression test to ensure the page does not fatal
for invalid parameter value.

Bug: T379293
Bug: T379291
Bug: T363408
Bug: T363486
Bug: T371621
Bug: T342358
Change-Id: I5a84bad9557668ad18ec6c09fbf1d926e50fa8d5