Forward backend warnings to Special:MediaSearch
The search api makes use of api warnings to inform the user about
problems executing the query, but Special:MediaSearch was not
presenting them to the user. Plumb the warnings through from
search responses into vue state, and render a basic warning box
containing the message.

The display of the warning is not the most elegant, but should
work for purpose. There was already a separate error display
for when no results are returned that can contain an error
message from the backend, but the warnings need to be more
general. They need to also be presented when there are results,
as they may describe limitations of the way the query was
executed.

Bug: T391876
Change-Id: I07e1fb235a7f1fc1fdba45a84601451518d17e95